perbaikan

+++ resources/views/admin/customers/components/actions.blade.php
@@ -1,5 +1,7 @@
 <div class="btn-group">
     <a href="{{ route('customers.show', $customer->id) }}" class="btn btn-sm btn-primary">Data</a>
-    <button class="btn btn-sm btn-warning" onclick="editCustomer({{ $customer->id }})">Edit</button>
-    <button class="btn btn-sm btn-danger " onclick="deleteCustomers({{ $customer->id }})">Hapus</button>
+    @if (Auth::user()->role != 'Owner')
+        <button class="btn btn-sm btn-warning" onclick="editCustomer({{ $customer->id }})">Edit</button>
+        <button class="btn btn-sm btn-danger " onclick="deleteCustomers({{ $customer->id }})">Hapus</button>
+    @endif
 </div>
