+++ resources/views/admin/stok/components/product/actions_product.blade.php
@@ -1,5 +1,7 @@
 <div class="btn-group">
     <a href="{{ route('products.show', $product->id) }}" class="btn btn-sm btn-primary">Rincian</a>
-    <button class="btn btn-sm btn-warning" onclick="editProduct({{ $product->id }})">Edit</button>
-    <button class="btn btn-sm btn-danger " onclick="deleteProduct({{ $product->id }})">Hapus</button>
+    @if (Auth::user()->role != 'Owner')
+        <button class="btn btn-sm btn-warning" onclick="editProduct({{ $product->id }})">Edit</button>
+        <button class="btn btn-sm btn-danger " onclick="deleteProduct({{ $product->id }})">Hapus</button>
+    @endif
 </div>
