+++ resources/views/admin/wirehouse/components/actions.blade.php
@@ -1,5 +1,7 @@
 <div class="btn-group">
     <a href="{{ route('wirehouses.show', $wirehouse->id) }}" class="btn btn-sm btn-primary">Data</a>
-    <button class="btn btn-sm btn-warning" onclick="editWirehouse({{ $wirehouse->id }})">Edit</button>
-    <button class="btn btn-sm btn-danger " onclick="deleteWirehouse({{ $wirehouse->id }})">Hapus</button>
+    @if (Auth::user()->role != 'Owner')
+        <button class="btn btn-sm btn-warning" onclick="editWirehouse({{ $wirehouse->id }})">Edit</button>
+        <button class="btn btn-sm btn-danger " onclick="deleteWirehouse({{ $wirehouse->id }})">Hapus</button>
+    @endif
 </div>
